Added box collision outline
When ran in debug mode you can see an outline for the box colliders to help with debugging. Almost works, seems to disappear when over a certain part of the screen, will make camera class so camera can move to help debug this. Also added "scene" variables like light position and colour.