Sound fall off and sound debug visuals
Audio files have to be mono for 3D sound due to OpenAL. Sound doesn't seem to be unhearable if out of the max distance range so I did it myself, sudden cut off to sound though when something is moving away. Also added spheres to visualise the max distance and the reference distance (how far you have to be for the sound to not be the max volume).