Added text to GUI
May change further. Currently only able to specify position first letter is drawn. Have got code which takes the desired position of the centre of the text and draws it centred, so may add that. Would also then be useful for the images and buttons to be drawn centred as well.