Added model collision
Lots of help from a friend of mine who lives in the computer. But yeah added a model collider. It gets all the triangles of the model and tests them for a collision with the other colliders. I did do model to model collision but it is very slow (predictably). For my final year project I won't need model to model collisions so I'll be fine. It runs faster if you collide with one of the triangles close to the start in the order they were defined in the .obj, since that is also the order they are written into the code and we can exit early. Will definitely be implementing a bounding box hierarchy for models to split the triangles up. How, I'm not sure yet. Maybe GetFaces() takes in a box that covers your collider, then GetFaces() only returns the faces that are in bounding boxes that the over arching box has collided with. That's the idea, but may not be the best way to architect it. Also added a maths helper with it's Maths namespace to put commonly used functions in, maybe like ray intercept if I end up adding rays. Big yap.