BVH for triangles in a model in detecting collision
Implemented bounding volume hierarchy for models. Triangles get split into nodes. Nodes have a min and max which can be used to check if a box you passed in (which should cover the whole of the other collision object) collides, if it does it checks for a collision with it's 2 children nodes. Either returns nothing if a collision between boxes isn't found, or returns triangles if a node that is being collided with doesn't have any children (leaf node). Easy concept, complicated code, the guy helped me out. Also decided to remove model to model collision. May add it back if I feel it is necessary in the future, but I don't foresee it being for my final year project. Currently can become slow at certain angles where I assume more nodes are being collided with, so more triangle tests. I found it isn't as slow if kludge is removed (kludge repeatedly moves the rigidbody little by little and checks if it's still colliding until it isn't anymore), I will be implementing physics based collision responses so collision detection only happens once which will be fine.