Added capsule collider
Think it works. I don't have a good debug model. I will go into uni tomorrow and try to make one on blender that is centred at 0,0,0. I'm not sure how to represent the dimensions as I have been using scale up to now, but the capsule is defined by a height and radius. If I change the z scale for radius I think the end cap spheres won't scale how I expect, but it might be fine.