Failed cylinder collision (again)
I can not get collision between a cylinder and a model to work. There is nothing online and I can't steer my friend who live in the computer to get it working correctly. I only need accurate collision on the curved face for a wheel. Thinking of using a library to do it for me.