Added ColDet for collision
Not lots of information online tbh but it seems to be able to give me the collision point from a triangle to triangle collision which is the main problem with my code at the moment. Was header only so can easily remove if it doesn't work well. Now need to figure out how to use it.