Added fcl
LONG day. Finally added fcl, to find the collision point from mesh to mesh collisions (all I wanted it for) isn't very accurate anyway. But it's not all bad because I can use the primitives provided for collision objects such as cylinders for wheels. I will have to rework my whole collision system and it means everything I did for it so far was pointless but it's fine. Also had to zip fcl.lib because it's too big to push. fcl without octomaps would be small enough but I couldn't get that to work.