Reworked collisions to use fcl entirely
Had to get rid of my BVH system. All in all lost minimum a weeks work, but probably more. It is for the greater good though as theoretically collisions will be better. Sphere to model seems good, box to model is a bit strange, but might be my physics. Also make sure to set convex to true and then set the model if your collision model is convex (recommended).