Added RayCollider (doesn't work)
Nothing is working how I wanted/expected it to. Decided to move to a ray to detect a collision for wheels against a track, turns out fcl doesn't have rays. Can use a long thin capsule instead acting as a ray, but fcl's non-convex mesh collision is very bad. It might be easier to revert everything I did this week and go back to before fcl. I could use my BVH and code my own ray code to get me a ray intersecting with a triangle. I would be able to get all triangles that potentially collide with the ray from my BVH structure and then I can carry out the ray-triangle intersect on each of those triangles. The contact point is apparently very accurate, so I could use that to see how close it is to the origin of the ray, and the max length of the ray minus that distance the contact point is to the origin should be the penetration depth. And I can simply get the normal from the triangle's normal. With this information I would simply do a rigidbody response on the entity with the ray collider and that should be good. I think I will revert my github back to before the fcl code. So this week I have made no (physical) progress 👍