Added suspension component
Not tested. Should work, probably doesn't. Decided to add a specific suspension class rather than a general spring class.

Car body should be one entity, then there are 4 anchor entities which are children of the car body. the anchor entities are where the suspension attaches to the car body. Suspension component can theoretically be added to any component as long as correct parameters set. Wheels are then not children of anything so their movement is individual. Will manually make them face the correct way with the model rotation offset which I also added to model renderer.