Made suspension better, fix rigidbody rotation
GETTING SOMEHERE! This whole time I wasn't using quaternions right in my rigidbody. May be the fix for things I was trying to fix before but moved on, so it's fine for now. Suspension does work! Getting the right values is very fiddly though. Made a makeshift car before doing the tire model. Although main problem right now is camera rotation. As soon as I try to get the camera to be on top of the car and facing the same direction, either by doing it manually or by making it a child of the car, there is a problem. As far as I can tell the problem is that it isn't rotating past 0-180 and just flipping back round. Probably to do with quaternions, very annoying.