Fixed suspension, added Tire component
Fixed the jitteriness in suspension by getting rid of lateral correction with force and instead correct the position manually overtime. This makes the car much more drivable, but if the car flips then it kind of breaks, probably fine, don't flip the car. With no tire forces the car is frictionless with the ground, wasn't on purpose but this is ideal for the tire model. I think I will manually set the wheels direction to be the same as the anchors (which are inherited from the car body), with the front tires adjusted for steering angle, and use that rotation in the tire component. I also plan to do the engine gears/throttle to rmp/rmp to torque manually, although an engine component may help with this. I also might decide to move everything into a "Car" component which you can add to an empty entity and all the stuff which I am going to do manually will be done for you. That would be after I get everything working. Hyped.