Did car textures, tried to fix suspension at high(er) weight
Suspension breaks at a higher (realistic) weight. Seems to only break at the start of the scene, if it makes it past that then it's fine. Also changed delta time at the start to 0 instead of simulating 60fps which seemed to help but the problem is still there. I think a fixed update would help this, but would require creating a thread. May have to be done.