Added wheels, corrected cars direction/proportions
Decided to leave suspension for today. The wheel with textures is in and they are working nicely. Car gets into a state where it's wheels are touching the ground and we're past max extension but it keeps pushing wheels into the ground and the car body up, which shouldn't be happening. Maybe redo suspension completely. Maybe faking it with non-rigidbody wheels and setting positions, but I really don't want to do that.