Tires working
Boom. Tires seem to work realistically. Pushing now as a milestone push, but may push again later. Tires rotate correctly and behave correctly from drive torque and brake torque. Right now brake torque is just torque in the opposite direction, like reversing. Need to change to make it so that it will only work to stop the tire and lock the tire if we lose traction. Also need to get controller support working so that you can actually play it.