Tire fiddling
Tire is constantly sliding left so applying force to counter the slide, then sliding right the next frame and countering that again. This is causing it to be very jittery. I think it is also using up all the "grip" so that might explain the overall lack of grip. If I can get that fixed then that would be great. It is only laterally, but the tires get up to +- 4 m/s and end up reaching the friction limit just standing "still".