WE'RE SO BACK
THIS WHOLE TIME the inertia tensor wasn't using an inertia tensor calculated with the actual weight, it was using one calculated from the default weight which was 1. That is why the suspension was so unstable and why the tires were so unstable because any force would move the object as if it was heavy but apply rotation torque as if it was very light. What a day. I only had to move the code that updates the inertia tensor to OnAlive (first frame of game) from the OnInitialise (as soon as it's made, before values can be set). I have increased the weight to the actual weight (1230kg) and increased the suspension forces with it and it still works. This means I can also increase the tire stiffness values and I no longer get any jittering. I can increase them to silly amounts and I don't get jittering. wowowow. Tire stuff may be done, apart from adjusting the values. Jubilations. I still need to limit the torque so it doesn't just increase forever actually but still. After that I need to get controller inputs working, get an actual track, get a timer I guess. Movin' on up!