Added controller support, and display inputs on screen
Added controller support through the input class with SDL. Implemented the controller input to the car controller, is cool. Also added a blend image GUI option which allows you to pass a blend factor in to have the GUI image be half one image and half the other. Blend seems like it's just mixing the 2 together though, like it the whole image into of the other one to the ratio you ask for, so maybe think of a different name. Anyway, used that to display throttle, brake, and steering input onto the screen, which is also cool.