Fixed flipping problem
The car would flip over on slopes. Turns out it is because the car would hit the ground, even though it doesn't look like it. Turned off collisions with car body and it is now fine, however car doesn't collide with any walls. A small price to pay.
Also fiddled with tire parameters. Still feels like we are always on the edge of grip, but that might just be getting realistic.
Also our top speed right now is 90kph while the real cars top speed is 250kph, so that isn't very realistic.