Added lap timer
Also fixed compression so that it is actually compression. Tried fixing wheels from jittering around but couldn't figure out exactly why. It is the whole wheel moving around because the anchor is moving around independent of the car body for some reason, even though it is anchored so it shouldn't. The whole wheel moving means the suspension (only for like 1 frame occasionally) doesn't give an upwards force. Not a massive problem in itself, but would be nice to make it not do that. Also visually it is a massive problem.