Implemented RPM/gears, added drag and tire rolling resistance
Implemented variable torque based off of gear ratios and RPM, and implemented RPM which is based off of wheel rotational speed. Seems to go up fairly quickly, don't know how realistic it is. Also gave very high torques sometimes to the point of losing control on a straight, which has been a reoccurring theme tbh. I limit the engine torque to 750Nm now before converting it to wheel torque. To try and help with losing control on a straight I increased longitudinal stiffness and decreased lateral stiffness to try and stop the overall grip exceeding the grip limit (gamma exceeding Fmax). Doesn't actually feel awful in the corners tbh, but I'm not sure if we are accelerating too quickly, or if that is down to too much lateral grip or not enough overall grip or too much wheel torque. Either way it's still very easy to spin out in 4th, which I don't think should be the case.
Also implemented drag and rolling resistance to limit the top speed/stop accelerating as fast. These might be the last forces I add(?) but I may increase them arteficially to try and slow it down a bit. On this track we never get to the max speed because it is so small irl, but also the model is no where near to scale. I might just go into Blender and scale it up by 2.