Fixed engine torque calculation
Apparently horsepower isn't constant, it changes based on the RPM. I now define a peak horsepower and we use a curve to pick the horsepower that should be used. So at low RPM we now have a very low horsepower so you accelerate very slowly, which is good. Also now don't allow any engine torque is you are at the rev limit, before the start was to stay in 4th the whole time because you could accelerate fast with no downside, but now you will stop accelerating when you hit the rev limit. Still wondering if something is wrong with the tire model, or I just need to find the right values. Hard to know where to put my time, into fiddling with the values or looking at the equations again. Either way it is still "slippery" when on a straight.