Lowered grip, added menu
Made it s that in tire it knows the default rest height of the suspension compress, and anything more compressed adds to the Fz (vertical load), this results in less overall grip than before but it's more realistic. Added a menu where you can change the input parameters, so people with more worn controllers can still play.