Let car change pitch due to cornering and acceleration
What I actually did is now apply the tire force too the car at the contact patch of the tire and the road instead of at the point the wheel connects to the car via suspension. This is more realistic, and purely from a programmatical point of view, each wheel provides more torque to the car since the force is further away from the centre of mass so the change is pitch is more pronounced. Now the outside tires have more grip while cornering due to the compressed suspension. However this introduced a bounce at high speed which was able to eventually flip the car. To counter this I have to change suspension forces a lot. This also means that now it feels more likely to flip you over due to curbs, which probably isn't good. If you don't touch a curb you'll be fine though, but you still shouldn't get flipped for coming off of a high curb too fast. Suspension is definitely the weak point.