Most recent version of the engine
Was making changes to the engine in a separate repository, shouldn't. I will continue development on Racing Game 3 in this repository so that this will have the most up to date version of the engine in. Then copy it over to the Racing Game 3 repository when I want to make a release. As a summary, overhauled suspension, now preparing to reoverhaul suspension by adding a raycast system. Also tire grip is determined by suspension forces rather than suspension compression, which is much more realistic.