Removed now unnecessary components from wheels
The RayCollider and Rigidbody have been removed from the wheels as neither of them are needed anymore. This does increase performance a fair amount as there are now less colliders and rigidbodies which before meant each rigidbody iterates over every collider. Next is to add a bump stop to stop the wheels going above the suspension limits.