Rolling resistance fix
When coasting the car didn't slow enough. This was because the tires were still attempting to deal with slip ratio generated by the added rolling resistance. Had to move code around so that we had the necessary variable where we needed them but that is fixed by limiting the longitudinal force to the net torque force (generated by the road torque, engine, or brake) so now when coasting the longitudinal force is 0 and the car slows down realistically. Had planned to adjust gear ratios and brake torque to align with ACC but this took up all my day.