Adjusted braking
Brakes now lock on their own through brake torque, not manually through arbitrary parameters. I first thought this was unrealistic but I hadn't developed the engine enough to realise that this was realistic. Also moved to using a maximum brake capacity which gets split to front/rear tires according to brake bias. Also increased the brake force to be more realistic, now easier to lock up because a higher brake force. Also attempted to make car behave more realistically. It does behave how it should with the values input, but getting the right values is difficult, specifically downforce, peak tire grip, and tire long and lat stiffness. Noticed a bug with suspension where it seems to keep the car at a certain height if it reaches that height (ie. specifically after aqua minirale going over the crest), like the ride height is set to a higher value, although I know it isn't being set, so maybe something to do with the maths. Also still need to adjust gear ratios to make the more realistic.