New suspension values
Also removed mSuspensionLength in Suspension because I'm not setting it and it doesn't need to be used, was responsible for the weird behaviour of the suspension "standing up" a few commits ago. But mainly played around with suspension values and various other things to get the feel right. Still seems like a huge lack of front grip, but as I said maybe it's realistic. I think the main front grip loss came when I implemented the separated damping for low/high bump/rebound, but I don't see why that would cause that. It's probably fine, isn't awful, but would be nice if you oversteered on entry if going too fast and applying too much steering rather than always understeering. Suspension set up curtesy of chatgpt as I'm not knowledgeable enough, had problems with bouncing because of the downforce on the straight but I guess that would happen with the wrong values so it's realistic. Also got suggested to raise the downforce levels, again it's very hard to tell what's realistic. The mix of downforce and peak tire grip is very difficult. Heavily feeling the effects of not having solid values that I know are true to the real world, doubting if things are working realistically for those values or if the system itself is broken.