Went back to directional lighting
Also prepared for implementing shadows. Previously had lighting only with point lights because I thought it would be easier. Point lights are actually much more difficult to implement shadows with, although admittedly much more flexible in the context of a game engine. Went back to directional light to add shadows. The lighting actually looks a lot better with directional light. Created a depth only render texture to use when rending the scene from an orthographic camera view from the directional lights direction. This will be used to sample from to see if a point is in light. Was going to use a camera component on an entity but now I am in two minds, because this would mean there'd be an extra entity created and "maintained" by the engine when I feel that there shouldn't be any "secret" entities that the user doesn't know about. May just not use the camera class and just do the maths "manually".