Implemented directional shadows
What a day! Probably the biggest single "thing" I've done, at least in terms of coding. In summary, scene first draws depth data to a render texture from the directional lights direction, then the shader uses that data to decide if each fragment gets the directional light. Was a big pain to implement, all coming down to not uploading the model's matrix in the new shadow shader. Probably a solid 7 hours straight worked on this, very cool. Also moved to the latest stable release of SDL2 because I updated NVIDIA drivers recently and the game controller SDL stuff was taking 30 seconds to detect a controller connecting/disconnecting, which obviously wasn't usable. That did fix itself when I turned my computer off and on again, so idk. Maybe some other stuff changed too but I don't even know at this point. Shadows do make it run a fair bit slower, could be decreased by making the resolution smaller so the render texture doesn't take as long to draw to, however I need the directional lights camera to cover a large area or the shadows popping in are very obvious/distracting, so you need a large resolution so that the shadows aren't so pixelated. I went from 700fps to around 350, so pretty expensive, but it needs to be above ~30 to not slow down the physics engine so a way to go, although a usable fps is probably 60, preferably 120, and my pc is decent, so should definitely keep optimisations in mind.