Moved Tire and Suspension demo, also thought
Makes more sense for Tire and Suspension to be inside demo instead of the actual engine. Reason they were there originally was because I wanted to make an engine for a driving game with the driving stuff done inside the engine, but it just makes more sense for the engine to do engine stuff and gameplay (like driving) done "by the user", or in the demo project. Also thought about the shadows more. Tried getting them to be a smooth line at the edge instead of pixilated due to shadow map resolution, but not really possible. Best thing to do is have multiple shadow maps, one high resolution but small and near the camera, then others getting less resolution and further from the camera. Could allow for less resolution overall in shadow maps, although have to loop through every model and ask it to draw for each shadow map. Again tricky part is just getting the required data nicely, but I think that is something I will definitely do.