Tried testing performance, weird
Added scoped timer to start when created and output how long it was alive for when it goes out of scope. The scene runs at the same fps in all these tests, 200, or 5ms per frame. If I create a scoped timer at the start of the RenderScene loop it gives me 5ms, which gives no time for anything else to be happening to achieve 200fps. If I then (only) add scoped timers inside the function measuring drawing to render textures and drawing the final scene, only leaving out a few opengl state sets and some light maths, each render texture takes 0.2ms and the final scene takes 0.3ms, but the overall time for the RenderScene function now says it takes 1.6ms. So I don't know what to believe anymore.