Added multiple prebaked shadow maps
Cool and works, but not how I wanted it to. I wanted to have multiple shadow maps (4) covering the same area so that the visual resolution of the shadows could be better. We were limited to the maximum size of a texture by the GPU (32k for me) and that was too stretched over the whole track resulting in a low visual resolution. I tried to have 4 textures at 32k split over the track so each texture covered less area and the visual resolution is better, but this (of course) results in lots of VRAM getting used to the point where my GPU (8GB) can't handle more than one 32k texture. So I had to settle with 4 textures of 16k, which is the same visual quality as 1 texture of 32k. So downside is that it was all for nothing but upside is that the way the GPU handles textures and looking up texture data means multiple smaller textures actually results in faster lookups, so it did increase performance. Also added detecting how much VRAM the GPU has and scaling the shadow map texture accordingly, only because I remembered my friend who tests it has less VRAM and it would crash for him. Also added 5x5 PCF in the shader and an early exit if a fragment isn't in shadow. Went from maybe 400fps without any oof the stuff added today to 450.