Added clutch
Was going to move the engine into it's own class but decided to do the clutch first. Basic clutch wasn't too difficult but the auto-clutch (so that I can actually drive with a controller) was much more difficult. Changing gears was fine (although no penalty if you don't use the clutch, maybe change so can't shift if clutch isn't in). Launching from a stand still was much more difficult. Clutch needs to be in at low revs to stop stalling (stalling not implemented but pretend) and ideally you build revs while clutch is in then release clutch and those revs get sent to the wheels, effectively avoiding the low revs. Turns out this is much more difficult to code, but I have something close. Also have a clutch with a biting point at 0.3 (0 being fully in, 1 fully out). Was able to get rid of the artificial torque boost at low rpm in first gear too now that a working clutch is in, so that is good. Still not entirely happy with how the auto-clutch works but may look into what exactly other games do.