Improved clutch
Did a lot more than that too. Firstly, clutch has an actual biting point at 0.35 to 0.65, torque is let into the tires linearly between these. Revs also get affected by the biting point, raising/lowering accordingly. Automatic clutch for moving off from a stand still improved. Also remapped the torque values for the engine. I was unaware but they were terrible before. I now use pairs to map an rpm to torque for a much more accurate curve, assuming you know what you're looking for. Now the peak torque is at around 5000rpm and the ideal shift rpm is around 7250, although it doesn't actually make a massive difference which I have seen with ACC. This in turn makes you accelerate a lot slower so the top speed became a lot lower due to lower torque for longer, so the drag coefficient was decreased to a now realistic value which is nice. Then also tried to fiddle around with the tire stiffness values as there was no tire spinning. I knew that it was very difficult to spin the tires when going around the track but there was tire spinning at the start due to artificially increased torque at low gears in first so that you could get away easily. Now I have removed that even at a stand still stamping the throttle you don't spin the rear tires. I think it could be due to my simplistic tire model, so I will work to expand the tire model and make it less "grip or no grip" and hope that makes the tires spin more. Or maybe I'll wait a little, do some other changes, make a release, then change the tire model and make that it's own release as it is a fairly big deal. Idk.