Optimised shader uniform uploading
Made it so that you (only "you" if "you" have access to the engine) can upload shaders once if they will always be the same or once perframe if they could change every frame but will be the same for every entity. I thought I would see a bigger increase in performance tbh but it seems kind of negligible. Either way definitely better as less data being uploaded to the GPU (same amount of data just less often).