Implemented poisson disk shadow sampling
PCF now used poisson disk. Basically, instead of the samples being in a grid, which maintains the pixel resolution of the shadow map due to the offsets of the samples being the size of the pixels, shadows now calculated using "random" (actually pre hard coded pseudo random samples placed to cover a good area) samples that are in a circle and rotated randomly (abusing overflow errors) for each fragment. This greatly breaks up the grid and makes a much smoother transition between shadow and no shadow. Also added some optimisations so that we can early exit if the pixel is/isn't probably in full shadow and then again we can early exit mid calculation if there are no changes to the result.