Theoretically added .glTF and .glb support
Not tested, but it theoretically works. Files are read in using the tiny_gltf library. Main reason was to set up PBR. Theoretically I can use glb files so I don't have to direct every texture to the material manually, but more importantly normal textures and others are already in, something I was struggling to do with my previous set up. Also implemented basic PBR in the shader, would read the PBR values of the model's materials but for now just uses the values we state in the ModelRenderer. Although it may be broken because the visual effect isn't what I expected, too shiny. Also improved fps from 400 to 500 for free by not looping over the maximum number of point lights every pixel every frame when we didn't even have any. A slight error on my part, assumed it would stop looping once it reached the end of the lights array, but there's no reason it would stop. Got rid of point lights all together for now, may want to reimplement at some point, but probably not for a while, was 100 fps for completely free.