Added IBL
Quite a lot of work for a not very noticeable effect tbh. IBL (Image Based Lighting) is where the material has a reflection of the surroundings. I just use the skybox for the reflection, and I do it a very simplified way. You're meant to have a blurred skybox image and I think 2 other separate textures, but I just use the original skybox texture and get multiple points which are averaged for a makeshift blur of the reflection, based on how rough it is. As I said not very noticeable, but it is there. Next hopefully confirm .glTF/.glb files actually work, then implement the PBR with the extra textures as maps.