Converted all models to .glb
Every model is now a .glb. Also separated the brake disk and the steering wheel, so now the brake disk doesn't spin when the wheel does and the steering wheel moves with your steering inputs. Weirdly, the skybox stops working when either the brake disk or the wheel model is used, I don't know why yet.