Changed menu, added vsync option
Menu broken into 3 categories, car, graphics, and controller. Might have to do something about making menus, because they are awful. Maybe set up some sort of automatic placement. Also added vsync toggle option, so your GPU can run the game and not be at 100% the whole time. Added that to the middle of the screen, so if I want to add something else then I'd have to change it's position, annoying, hence why I should change it to not be like that. Also added button state enums to be returned instead of ints for the button states. And finally added last frame time, which is the same as delta time but it isn't affected by the time scale, so can be used when you actually want to track real life time separate from the delta time which would get paused if time scale was set to 0. I use it for frame rate calculations/interval between showing frame rate.