Added adjustable brake bias
Also added more downforce. As said many times it is so difficult to know how much downforce there actually should be. The values I have now seems to have multiple (more than one) sources saying these similar numbers, however how much to the front and rear and if this is in a high downforce set up (like you would expect at Imola), I have no idea. If we assume downforce is perfect now, we also have the overall grip of the tire to adjust and the individual longitudinal and lateral stiffness, the stiffness values being difficult to visualise the changes. And then I don't know if it is actually realistic or I just find it nice to drive too. Long day. But should just make it a release tomorrow.