Adjusted tire parameters
Specifically stiffness, and decreased it by a lot. Did a deep dive into the tires. Stiffness represents how quickly the tire reaches the peak force. So a higher stiffness means less slip angle (laterally) is needed before the tire is saturated with force. Simple to think about but visualising it into the handling is more difficult. Did some tests, found the general consensus for the peak slip angle is 6 degrees and the peak slip ratio is 30%, so adjusted my stiffness values to match that. I think because my tire model is just a simple "grip or no grip" it behaves weirdly with these low values. Temporarily increased them so that it is playable for the release, but will hopefully expand the tire model soon and be able to use the actual stiffness values. Also tried to make the audio not be directional if inside the minimum distance, so when sitting in the car the audio isn't only in the right ear, but it was difficult and possibly OpenAL isn't actually that good.