Glass now renders, also added double sided geometry support
Glass isn't implemented properly I think, but it renders now. Double sided support is a pain. There are many pieces of geometry that look like they're fighting over the z axis after enabling it. Need to fix, or change the model? It doesn't do it in Blender or any other 3D viewing software so I guess it is just me.