Split peak friction into long and lat
Was finding that my tires wouldn't spin up if I floored throttle from a stand still like they should (so need less grip), but I was taking a correct amount of speed through corners. Solution was to separate the 2 directions, so there is less grip longitudinally than laterally, which is realistic. Still very difficult to get actual values that are realistic though.