Implemented cascade shadows properly
Now runs slower, mainly because it draws the track in the cascade shadow maps which takes a while because the track is so big. Upside is the tracks shadows are much higher quality and we use much less vram, downside is performance. I went from 470fps with pre-baked shadow maps for the track, so the track didn't have to render in the cascade shadow maps, to 330fps with 3 cascade shadow maps, which is painful. But now the cascades fill up the view more efficiently so they're theoretically getting a higher visual resolution. Also made a new track model where I got rid of the ground which isn't needed for shadows so the rendering of the shadow maps should be more efficient.