Moved to 5 point collision
Like ACC. So there are now 5 rays that detect collision instead of 1, and an average is found to get the position/surface normal. Also found out that I was still increasing the inertia value for the tires from WAY back when my suspension couldn't handle the car being too heavy (was a skill issue) so I decreased all the weights by 10. I increased the inertia by 10 in an attempt to compensate. Anyway, I fixed that issue and started using the actual weights, but I obviously forgot to reduce the tire inertia again. So this whole time the tire was 10 times more difficult to make roll. Fixing it has opened a whole world of problems, which will be fixed tomorrow.