Implemented closed-form backwrads-euler tire rotation logic
Problems seems to be from the integration. Previous integration wasn't sophisticated enough, and it seems neither is this one. Will attempt to implement a better one tomorrow. This one doesn't have the true effect of the drive and brake torque, and does a weird thing of setting the drive torque to go backwards when speed/drive torque is low.