Implemented fully-implicit wheel integration
It was all the guy in the screen icl. Had an issue with the wheel jittering at low speed/stationary. All fixed. Pretty sure it uses a "if low speed do this" which I don't like but it works well enough it seems. Still not happy with how brake torque works because it can reverse the tire direction and then cause it to flip flop frame to frame.