Added dynamic contact patch
Also added a chase cam that lags behind the direction of velocity. Should implement a "LookAt()" function in Transform. On another note the lap times are fairly close to what I got in ACC again. Would be good to get different amount of grip for the surface the tire is on, I could image that is finding what material the triangle of the track model is, could be expensive.