Added drivetrain which has differential and torque splitting
I'm back. Implemented a differential system as well as torque splitting through a drivetrain. Rear wheels are somewhat locked together so that they don't spin freely anymore, and more torque is given to outside wheels to aid rotation (happens in real life). To be honest I can't tell much of a difference when I drive it. Also tweaked the auto clutch behaviour so that if you lock the rear wheels and the clutch engages to stop a "stall" then you no longer need to apply some throttle to get the clutch to disengage.