Decreased longitudinal grip, adjusted feedback
Tire now feels closer to the grip levels in ACC where tires spin while accelerating fully from first or second and still maintain speed we want through corners. Also changed how the tire screech for tire spin and locking (slip ratio) is calculated, while now using slip angle for the controller vibration. Both of these values are just "plucked out of thin air", it would be good to get them based on the peak slip angle/ratio at that moment so we have an accurate idea of if we are abusing the tires or not. Also tested the drivetrain stuff with different parameters and they do make a difference, some more than others.