Added file read/write, added saved lap time
Can now read/write to a file using functions in Core. Used that to save your fastest lap as well as the sample data for the delta time. Would be cool to export telemetry (throttle/brake input, speed) and be able to compare laps, but they're built on it being over a distance which I don't know how you actually calculate.