Started rendering overhaul
To confirm, overhaul of how rendering works, not actual graphics, they will look the same but hopefully the new method allows for more advanced techniques. Instead of relying on the "immediate" ModelRenderer rendering the model itself, which we can't guarantee the order of and we have no idea of any other models in the scene, we'll pass our model to the SceneRenderer which will be able to render all opaque materials last, as well as many other techniques. This is also better putting all rendering code in one place. I will also move the lighting stuff into there as well. I am also thinking about changing the shader to be a "use()" rather than using and unusing the shader every time we call (so doing it like on Learn OpenGL) so that the decreased calls can help with frame rate. Exciting times.